use more fitting integer types for function result

+++ libc-bottom-half/sources/wasip3_file_utils.c
@@ -3,8 +3,8 @@
 #ifdef __wasip3__
 #include <wasi/file_utils.h>
 
-ssize_t __wasilibc_write_stream3(int fildes, wasip3_write_t **write_end,
-                                 off_t **off) {
+int __wasilibc_write_stream3(int fildes, wasip3_write_t **write_end,
+                             off_t **off) {
   descriptor_table_entry_t *entry = descriptor_table_get_ref(fildes);
   if (!entry)
     return -1;
@@ -15,7 +15,7 @@ ssize_t __wasilibc_write_stream3(int fildes, wasip3_write_t **write_end,
   return (*entry->vtable->get_write_stream3)(entry->data, write_end, off);
 }
 
-ssize_t __wasilibc_read_stream3(
+int __wasilibc_read_stream3(
     int fildes,
     filesystem_tuple2_stream_u8_future_result_void_error_code_t **stream,
     off_t **off) {
